% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedit.R
\docType{class}
\name{GEdit-class}
\alias{GEdit}
\alias{GEdit-class}
\title{show the word list}
\arguments{
\item{str}{a string. If
missing do nothing, otherwise match against
string to generate word list. Popup menu
depending on length}
}
\description{
show the word list
}
\section{Methods}{

\describe{
\item{\code{add_bindings()}}{Add bindings to the entry box}

\item{\code{clear_error()}}{Clear error message}

\item{\code{clear_init_txt(...)}}{clear out init text, set back to black}

\item{\code{find_match(x)}}{Find match in word list}

\item{\code{make_styles(bg = "#ff6622")}}{Create tcl styles, cf http://paste.tclers.tk/506}

\item{\code{set_error(msg)}}{Add error state and message to widget}

\item{\code{set_init_txt(msg)}}{set initial text, gray out}

\item{\code{set_invalid(value, msg)}}{Set widget as invalid or not}

\item{\code{set_validator(FUN)}}{Set a function to do the validation}

\item{\code{validate_input()}}{Return logical indicating if input is valid}
}}

