% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk-misc.R
\name{getWidget.tkwin}
\alias{getWidget.tkwin}
\title{method for getWidget defined in gWidgets2}
\usage{
\method{getWidget}{tkwin}(obj)
}
\arguments{
\item{obj}{object}
}
\description{
method for getWidget defined in gWidgets2
}

