\name{GStatusBar}
\alias{GStatusBar}
\title{\code{GStatusBar} is the base class for a status bar}
\description{
  The \code{GStatusBar} class inherits for \code{GBoxContainer} meaning it can be used as a parent container. As such, one can add additional widgets beyond the plain label that is the main property of this widget.
}

