% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoolbar.R
\docType{class}
\name{GToolBar-class}
\alias{GToolBar}
\alias{GToolBar-class}
\title{\code{GToolBar} is the base class for toolbars}
\arguments{
\item{...}{passed to constructor}
}
\description{
The toolbar is a container, so can have other widgets added to it
as though it were a box container. Buttons should be added as
action items, so that they are rendered in the proper style. Check
buttons should be given the argument \code{use.togglebutton}. Use
\code{addSpring} to right align items.
}
\section{Methods}{

\describe{
\item{\code{add_gaction_toolitem(obj)}}{Helper to add a gaction item}

\item{\code{add_gseparator_toolitem()}}{Helper to add a separator}

\item{\code{add_toolbar_items(items)}}{Map a toolbar list, a named list of gaction items or gsepartor items}

\item{\code{clear_toolbar()}}{Clear toolbar items}

\item{\code{get_widget()}}{What widget do we use for the parent of the children}

\item{\code{set_value(value, ...)}}{We can't really adjust spacing between children after they have been positioned.}
}}

