\name{.gexpandgroup.guiWidgetsToolkitRGtk2}
\alias{.gexpandgroup.guiWidgetsToolkitRGtk2}
\title{toolkit constructor}
\usage{
  \method{.gexpandgroup}{guiWidgetsToolkitRGtk2} (toolkit,
  text, markup, horizontal = TRUE, handler = NULL, action =
  NULL, container = NULL, ...)
}
\description{
  toolkit constructor
}

