% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendar.R
\name{makeCalendar}
\alias{makeCalendar}
\title{make a calendar...}
\usage{
makeCalendar(date_var, widget, date, date_format = "\%Y-\%m-\%d", set_value)
}
\arguments{
\item{date_var}{var}

\item{widget}{var}

\item{date}{var}

\item{date_format}{var}

\item{set_value}{var
from chron with slight change to arguments}
}
\description{
make a calendar...
}

