\name{GSpinButton}
\alias{GSpinButton}
\title{Spinbutton class}
\usage{
  GSpinButton(...)
}
\arguments{
  \item{...}{passed to constructor \code{GSpinButton} is
  the base class for spin buttons. The widget is buggy in
  tcltk, atleast with the old non-themed style and a Mac
  running tcltk under X11. The bug is the thing keeps
  spinning when the buttons are pressed.}
}
\description{
  Spinbutton class
}

