% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.R
\name{gwidgets2_tcltk_column_alignment}
\alias{gwidgets2_tcltk_column_alignment}
\alias{gwidgets2_tcltk_column_alignment.default}
\alias{gwidgets2_tcltk_column_alignment.numeric}
\alias{gwidgets2_tcltk_column_alignment.logical}
\title{align a column based on the class of the content}
\usage{
gwidgets2_tcltk_column_alignment(x)

\method{gwidgets2_tcltk_column_alignment}{default}(x)

\method{gwidgets2_tcltk_column_alignment}{numeric}(x)

\method{gwidgets2_tcltk_column_alignment}{logical}(x)
}
\arguments{
\item{x}{column data to align}
}
\value{
anchor string
}
\description{
Gives visual difference to data based on its class
}
