\name{List-class}
\alias{List}
\alias{List-class}
\title{A list extension class.}
\usage{
  List(...)
}
\arguments{
  \item{...}{passed to constructor}
}
\description{
  Like a list, but has some methods. Completely superflous,
  but makes copying some code algorithms easier. We
  implement methods such as \code{append}, \code{push},
  \code{pop} and \code{each} for iteration. As well, there
  are some lookup methods.
}

